+++ demisto_sdk/commands/common/hook_validations/xdrc_templates.py
@@ -7,9 +7,9 @@ from demisto_sdk.commands.common.hook_validations.base_validator import error_co
 from demisto_sdk.commands.common.hook_validations.content_entity_validator import ContentEntityValidator
 
 
-class XSIAMReportValidator(ContentEntityValidator):
+class XDRCTemplatesValidator(ContentEntityValidator):
     """
-    XSIAMReportValidator is designed to validate the correctness of the file structure we enter to content repo.
+    XSIAMDashboardValidator is designed to validate the correctness of the file structure we enter to content repo.
     """
 
     def __init__(self, structure_validator, ignored_errors=None, print_as_warnings=False, json_file_path=None):
@@ -19,7 +19,7 @@ class XSIAMReportValidator(ContentEntityValidator):
 
     def is_valid_file(self, validate_rn=True, is_new_file=False, use_git=False):
         """
-        Check whether the xsiam report is valid or not
+        Check whether the xsiam dashboard is valid or not
         Note: For now we return True regardless of the item content. More info:
         https://github.com/demisto/etc/issues/48151#issuecomment-1109660727
         """
@@ -32,13 +32,13 @@ class XSIAMReportValidator(ContentEntityValidator):
         """
         pass
 
-    @error_codes("XR100")
+    @error_codes("XT100")
     def is_files_naming_correct(self):
         """
         Validates all file naming is as convention.
         """
         if not self.validate_xsiam_content_item_title(self.file_path):
-            error_message, error_code = Errors.xsiam_report_files_naming_error([self.file_path])
+            error_message, error_code = Errors.xdrc_templates_files_naming_error([self.file_path])
             if self.handle_error(error_message, error_code, file_path=self.file_path):
                 self._is_valid = False
                 return False