+++ demisto_sdk/commands/common/tools.py
@@ -30,6 +30,7 @@ from git.types import PathLike
 from packaging.version import LegacyVersion, Version, parse
 from pebble import ProcessFuture, ProcessPool
 from requests.exceptions import HTTPError
+from ruamel.yaml.comments import CommentedSeq
 
 from demisto_sdk.commands.common.constants import (ALL_FILES_VALIDATION_IGNORE_WHITELIST, API_MODULES_PACK,
                                                    CLASSIFIERS_DIR, DASHBOARDS_DIR, DEF_DOCKER, DEF_DOCKER_PWSH,
@@ -1475,7 +1476,7 @@ def find_type(
             if MODELING_RULES_DIR in Path(path).parts:
                 return FileType.MODELING_RULE
 
-        if 'global_rule_id' in _dict:
+        if 'global_rule_id' in _dict or (isinstance(_dict, CommentedSeq) and _dict and 'global_rule_id' in _dict[0]):
             return FileType.CORRELATION_RULE
 
     if file_type == 'json' or path.lower().endswith('.json'):